Refactor test module: `test_downloader`
