+++ tests/test_bbbike_d.py
@@ -1,0 +1,182 @@
+"""
+Test the :py:class:`pydriosm.downloader.bbbike.BBBikeDownloader` class.
+"""
+
+import os
+
+import pandas as pd
+import pytest
+from pydriosm.downloader.bbbike import BBBikeDownloader
+from pyhelpers.dirs import delete_dir
+
+
+class TestBBBikeDownloader:
+
+    @pytest.fixture(scope='class')
+    def bbd(self):
+        return BBBikeDownloader()
+
+    def test_init(self, bbd):
+        assert bbd.NAME == 'BBBike'
+        assert bbd.LONG_NAME == 'BBBike exports of OpenStreetMap data'
+        assert bbd.URL == 'https://download.bbbike.org/osm/bbbike/'
+        assert os.path.relpath(bbd.download_dir) == 'osm_data\\bbbike'
+
+        bbd_ = BBBikeDownloader(download_dir="tests\\osm_data")
+        assert os.path.relpath(bbd_.download_dir) == 'tests\\osm_data'
+
+        assert isinstance(bbd.valid_subregion_names, list)
+        assert isinstance(bbd.subregion_coordinates, pd.DataFrame)
+        assert isinstance(bbd.subregion_index, pd.DataFrame)
+        assert isinstance(bbd.catalogue, dict)
+
+    @pytest.mark.parametrize('update', [True, False])
+    def test_get_names_of_cities(self, bbd, update):
+        bbbike_cities = bbd.get_names_of_cities(
+            update=update, confirmation_required=False, verbose=True)
+        assert isinstance(bbbike_cities, list)
+
+    @pytest.mark.parametrize('update', [True, False])
+    def test_get_coordinates_of_cities(self, bbd, update):
+        coords_of_cities = bbd.get_coordinates_of_cities(
+            update=update, confirmation_required=False, verbose=True)
+
+        assert isinstance(coords_of_cities, pd.DataFrame)
+        assert coords_of_cities.columns.to_list() == [
+            'city',
+            'real_name',
+            'pref._language',
+            'local_language',
+            'country',
+            'area_or_continent',
+            'population',
+            'step',
+            'other_cities',
+            'll_longitude',
+            'll_latitude',
+            'ur_longitude',
+            'ur_latitude']
+
+    @pytest.mark.parametrize('update', [True, False])
+    def test_get_subregion_index(self, bbd, update):
+        subrgn_idx = bbd.get_subregion_index(update=update, confirmation_required=False, verbose=True)
+
+        assert isinstance(subrgn_idx, pd.DataFrame)
+        assert subrgn_idx.columns.to_list() == ['name', 'last_modified', 'url']
+
+    def test_get_valid_subregion_names(self, bbd):
+        subrgn_names = bbd.get_valid_subregion_names()
+
+        assert isinstance(subrgn_names, list)
+
+    def test_validate_subregion_name(self, bbd):
+        subrgn_name = 'birmingham'
+
+        valid_name = bbd.validate_subregion_name(subregion_name=subrgn_name)
+        assert valid_name == 'Birmingham'
+
+    def test_get_subregion_catalogue(self, bbd, capfd):
+        subrgn_name = 'birmingham'
+
+        bham_dwnld_cat = bbd.get_subregion_catalogue(
+            subregion_name=subrgn_name, confirmation_required=False, verbose=True)
+        out, _ = capfd.readouterr()
+        assert 'Compiling the data of a download catalogue for "Birmingham" ... Done.\n' == out
+        assert isinstance(bham_dwnld_cat, pd.DataFrame)
+        assert bham_dwnld_cat.columns.to_list() == [
+            'filename', 'url', 'data_type', 'size', 'last_update']
+
+    def test_get_catalogue(self, bbd):
+        bbbike_catalogue = bbd.get_catalogue()
+        assert list(bbbike_catalogue.keys()) == ['FileFormat', 'DataType', 'Catalogue']
+
+        catalogue = bbbike_catalogue['Catalogue']
+        assert isinstance(catalogue, dict)
+
+        bham_catalogue = catalogue['Birmingham']
+        assert isinstance(bham_catalogue, pd.DataFrame)
+
+    def test_validate_file_format(self, bbd):
+        valid_file_format = bbd.validate_file_format(osm_file_format='PBF')
+        assert valid_file_format == '.pbf'
+
+        valid_file_format = bbd.validate_file_format(osm_file_format='.osm.pbf')
+        assert valid_file_format == '.pbf'
+
+    def test_get_subregion_download_url(self, bbd):
+        subrgn_name = 'birmingham'
+        file_format = "pbf"
+
+        subrgn_name_, dwnld_url = bbd.get_subregion_download_url(subrgn_name, file_format)
+        assert subrgn_name_ == 'Birmingham'
+        assert dwnld_url == 'https://download.bbbike.org/osm/bbbike/Birmingham/Birmingham.osm.pbf'
+
+        file_format = "csv.xz"
+        subrgn_name_, dwnld_url = bbd.get_subregion_download_url(subrgn_name, file_format)
+
+        assert subrgn_name_ == 'Birmingham'
+        assert dwnld_url == \
+               'https://download.bbbike.org/osm/bbbike/Birmingham/Birmingham.osm.csv.xz'
+
+    def test_get_valid_download_info(self, bbd):
+        subrgn_name = 'birmingham'
+        file_format = "pbf"
+
+        info = bbd.get_valid_download_info(subrgn_name, file_format)
+        valid_subrgn_name, pbf_filename, dwnld_url, pbf_pathname = info
+
+        assert valid_subrgn_name == 'Birmingham'
+        assert pbf_filename == 'Birmingham.osm.pbf'
+        assert dwnld_url == 'https://download.bbbike.org/osm/bbbike/Birmingham/Birmingham.osm.pbf'
+        assert os.path.relpath(pbf_pathname) == 'osm_data\\bbbike\\birmingham\\Birmingham.osm.pbf'
+
+        bbd_ = BBBikeDownloader(download_dir="tests\\osm_data")
+        _, _, _, pbf_pathname = bbd_.get_valid_download_info(subrgn_name, file_format)
+        assert os.path.relpath(pbf_pathname) == 'tests\\osm_data\\birmingham\\Birmingham.osm.pbf'
+
+    def test_file_exists(self, bbd):
+        subrgn_name = 'birmingham'
+        file_format = ".pbf"
+        dwnld_dir = "tests\\osm_data"
+
+        pbf_exists = bbd.file_exists(subrgn_name, file_format, dwnld_dir)
+        assert not pbf_exists
+
+    # @staticmethod
+    # def test_download_subregion_data(capfd):
+    #     subrgn_name = 'leeds'
+    #     dwnld_dir = "tests\\osm_data"
+    #
+    #     bbd.download_subregion_data(
+    #         subregion_name=subrgn_name, download_dir=dwnld_dir, ret_download_path=True,
+    #         verbose=True, confirmation_required=False)
+    #     out, _ = capfd.readouterr()
+    #     assert 'Check out the downloaded OSM data at "tests\\osm_data\\leeds\\".' in out
+
+    def test_download_osm_data(self, bbd):
+        subrgn_name = 'London'
+        file_format = "pbf"
+
+        bbd.download_osm_data(subrgn_name, file_format, confirmation_required=False)
+        assert len(bbd.data_paths) == 1
+        assert os.path.relpath(bbd.data_paths[0]) == 'osm_data\\bbbike\\london\\London.osm.pbf'
+
+        london_dwnld_dir = os.path.relpath(bbd.download_dir)
+        assert london_dwnld_dir == 'osm_data\\bbbike'
+
+        subrgn_names = ['leeds', 'birmingham']
+        dwnld_dir = "tests\\osm_data"
+
+        dwnld_paths = bbd.download_osm_data(
+            subrgn_names, file_format, dwnld_dir, confirmation_required=False,
+            ret_download_path=True)
+        assert len(dwnld_paths) == 2
+        assert len(bbd.data_paths) == 3
+        assert os.path.relpath(bbd.download_dir) == os.path.relpath(dwnld_dir)
+        assert os.path.relpath(os.path.commonpath(dwnld_paths)) == 'tests\\osm_data'
+
+        delete_dir([os.path.dirname(london_dwnld_dir), dwnld_dir], confirmation_required=False)
+
+
+if __name__ == '__main__':
+    pytest.main()
